s3smash.py
This Python script performs an S3 upload/download test using multiple threads. It takes user input for object size, number of objects, and number of threads to use. The boto3 and threading libraries are used to perform the tests in parallel. The script records the time it takes to upload/download each object and calculates the throughput in megabytes per second.


A dictionary called results is initialized to store the test results. Two lists called upload_times and download_times are also initialized to store the upload and download times of each object.

There are three functions defined: upload_object, download_object, and calculate_throughput.

The upload_object function takes a thread_id as an argument and uploads a specified number of objects to S3 with a given object size. The object names are generated using the thread_id and object number. The function records the time it takes to upload each object and stores it in the results dictionary. 

The download_object function is similar to the upload_object function, but it downloads the objects instead.

The calculate_throughput function takes the upload and download times, object sizes, and calculates the throughput in megabytes per second.

The script creates a list of upload and download threads using the threading library and waits for all threads to finish before recording the total time it takes to complete the upload and download tests.

The script then prints the results of the test, including the time it took to upload/download each object, the average, minimum, and maximum upload/download times, and the average, minimum, and maximum upload/download throughput in megabytes per second.

I plot the upload and download throughput to a local PNG file using the matplotlib library.

Note:  I hard coded the S3 bucket name, you can easily prompt for this as well if you like. 

Note2:  Larger instance types have higher network throughput, so results will vary, this test harness provides a good way to profile Amazon S3 from a variety of instance types and from different regions.